+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Version 7.0.0-dev
 - BC break: removed BaseAttribute
 - fixed failure line in JUnit reports
 - BC break: replaced method IResultsFormatter::render with outputResults
+- BC break: only array can be passed to attribute Skip (if any value is passed)
 
 Version 6.1.0
 - added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter(), TestsStartedEvent, TestsFinishedEvent to public api